refactor(api): PR-6.5 commit 4 — files/curation/duplicates 12 routes
server.py 1354 → 957 行（−397），projects 域第四刀（文件管理 + curation +
去重）。

  api/routers/projects/curation.py    12 routes  files delete + list, project_thumb,
                                                 versions/{vid}/jobs/latest,
                                                 versions/{vid}/curation get +
                                                 copy/remove/folder,
                                                 preprocess/duplicates/scan+apply,
                                                 backward-alias /duplicates/scan+apply
                                                 + _curation_err_code +
                                                 _duplicate_err_code helpers
  api/schemas/curation.py             6 BaseModel  DeleteFilesRequest /
                                                   CopyRequest / RemoveRequest /
                                                   FolderOp /
                                                   DuplicateScanRequest /
                                                   DuplicateApplyRequest

回归：PR-1 安全网 152 + project_endpoints 49 = **207/207 全绿**。
route snapshot 不变（160 routes）。
